requirements.txt
akshare>=1.0.0
tushare>=1.0.0
yfinance>=0.2.0
pyyaml>=6.0
jinja2>=3.0
openclaw>=1.0.0
pytest>=7.0
pytest-cov>=4.0
